Fixed a bunch of bugs

+++ server/schema/typeDefs.js
@@ -14,7 +14,7 @@ type User{
  _id: ID
  username: String!
  email: String
- bookCount: Int16Array
+ bookCount: Int
  savedBooks: [Book]
 }
